Added drop chance to item after kill entity

+++ src/main/java/com/arturoo404/game/entity/EntityDropItem.java
@@ -11,6 +11,6 @@ import lombok.*;
 public class EntityDropItem {
 
     private ItemName itemName;
-    private int dropChance;
+    private double dropChance;
     private int dropQuantity;
 }
